Fix performance bug when using g for nodes. Add D/R labels to nodes
There was a preformance bug where each time a node was added with a g
group, every node was added again. This meant if you added n nodes,
there would be n svg:g groups, and within each of the groups there
would be n copies of the 2 text and circle elements that defined the
group.

I fixed the bug and added D and R labels. I think the new
visualization is easier to understand. I could never remember which
side was the donor and which side was the recipient before.

But I still find it strange that a donor edge that leaves a node
(recipient edge that enters respectively) can sometimes originate
from the recipient (donor) side. In other words the color of the
edge should match the color of side of the node from which that
edge enters or leaves.