Switch D and R in tables. Remove manual GUI
I wanted to put the donor in the rows and the recipient in the columns.
So a (donor, recipient) pairs corresponded to (i,j) pairs in a table.
But I couldn't understand the code that built the custom UI.

So I removed the custom UI and replaced it with a table with
input boxes.